+++ src/fides/api/ctl/routes/generate.py
@@ -10,13 +10,13 @@ from loguru import logger as log
 from pydantic import BaseModel, root_validator
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from fides.api.api.v1 import scope_registry
 from fides.api.ctl.database.crud import get_resource
 from fides.api.ctl.database.session import get_async_db
 from fides.api.ctl.routes.util import API_PREFIX
 from fides.api.ctl.sql_models import sql_model_map  # type: ignore[attr-defined]
 from fides.api.ctl.utils.api_router import APIRouter
-from fides.api.ops.api.v1 import scope_registry
-from fides.api.ops.oauth.utils import verify_oauth_client_prod
+from fides.api.oauth.utils import verify_oauth_client_prod
 from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,