+++ src/fides/api/ctl/routes/validate.py
@@ -7,10 +7,10 @@ from typing import Callable, Dict, Union
 from fastapi import Response, Security, status
 from pydantic import BaseModel
 
+from fides.api.api.v1 import scope_registry
 from fides.api.ctl.routes.util import API_PREFIX
 from fides.api.ctl.utils.api_router import APIRouter
-from fides.api.ops.api.v1 import scope_registry
-from fides.api.ops.oauth.utils import verify_oauth_client_prod
+from fides.api.oauth.utils import verify_oauth_client_prod
 from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,