+++ src/fides/api/service/connectors/email/attentive_connector.py
@@ -1,0 +1,7 @@
+from fides.api.service.connectors.erasure_email_connector import (
+    GenericErasureEmailConnector,
+)
+
+
+class AttentiveConnector(GenericErasureEmailConnector):
+    """No custom logic here but we want to consider this as a separate connector"""