+++ src/fides/cli/utils.py
@@ -24,7 +24,7 @@ from fideslog.sdk.python.utils import (
 from requests import get, put
 
 import fides
-from fides.api.ops.api.v1.urn_registry import REGISTRATION, V1_URL_PREFIX
+from fides.api.api.v1.urn_registry import REGISTRATION, V1_URL_PREFIX
 from fides.common.utils import check_response, echo_green, echo_red
 from fides.connectors.models import (
     AWSConfig,