+++ tests/ctl/core/test_custom_field_models.py
@@ -3,8 +3,8 @@
 import pytest
 from sqlalchemy.exc import IntegrityError
 
+from fides.api.common_exceptions import KeyOrNameAlreadyExists
 from fides.api.ctl.sql_models import CustomFieldDefinition
-from fides.api.ops.common_exceptions import KeyOrNameAlreadyExists
 
 
 @pytest.fixture(autouse=True)