+++ tests/ops/api/v1/endpoints/test_registration_endpoints.py
@@ -2,8 +2,8 @@ from unittest import mock
 
 import pytest
 
-from fides.api.ops.api.v1.urn_registry import REGISTRATION, V1_URL_PREFIX
-from fides.api.ops.models.registration import UserRegistration
+from fides.api.api.v1.urn_registry import REGISTRATION, V1_URL_PREFIX
+from fides.api.models.registration import UserRegistration
 
 
 class TestUserRegistration:
@@ -48,9 +48,7 @@ class TestUserRegistration:
         data = resp.json()
         assert data["opt_in"] == True
 
-    @mock.patch(
-        "fides.api.ops.api.v1.endpoints.registration_endpoints.send_registration"
-    )
+    @mock.patch("fides.api.api.v1.endpoints.registration_endpoints.send_registration")
     def test_register_user(
         self,
         send_registration_mock,
@@ -77,9 +75,7 @@ class TestUserRegistration:
         assert len(UserRegistration.all(db)) == 1
         assert send_registration_mock.call_count == 1
 
-    @mock.patch(
-        "fides.api.ops.api.v1.endpoints.registration_endpoints.send_registration"
-    )
+    @mock.patch("fides.api.api.v1.endpoints.registration_endpoints.send_registration")
     def test_register_user_upserts_locally(
         self,
         send_registration_mock,
@@ -121,9 +117,7 @@ class TestUserRegistration:
         assert resp.status_code == 422
         assert len(UserRegistration.all(db)) == 0
 
-    @mock.patch(
-        "fides.api.ops.api.v1.endpoints.registration_endpoints.send_registration"
-    )
+    @mock.patch("fides.api.api.v1.endpoints.registration_endpoints.send_registration")
     def test_register_user_one_allowed(
         self,
         send_registration_mock,